auth-sync-tool/pyproject.toml
[project]
name = "stateful-abac-auth-sync"
version = "0.1.0"
description = "CLI tool to generate Stateful ABAC Policy Engine manifests from external databases"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
stateful-abac-sync = "stateful_abac_sync.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stateful_abac_sync"]
